Create Analysis file
This is the code that contains  all the analysis I did over the weekend. Note: you might get error when you run the code because alot of the file pathways are for my computer not yours and so you might need to adjust that. Let me know if you find such errors. Thank :)